updated createthread for generatetitle
